# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter API",
    "description": "\nA clean, read-only REST API over the **Urja Meter Ops** portal (smart-meter data for a\nJaipur distribution utility). The portal has no API of its own. This service logs in as a\nnormal user, reads the portal's internal endpoints (see `PROTOCOL.md` in the repository)\nand serves the data as typed JSON.\n\n* **Meters**: nameplate, status, location, network position; filter by any attribute,\n  network node or distance.\n* **Readings / consumption**: register readings normalised to numbers and ISO timestamps\n  (IST), with per-interval and per-day consumption derived from register deltas.\n* **Network**: the zone \u2192 circle \u2192 division \u2192 subdivision \u2192 substation \u2192 feeder \u2192\n  transformer hierarchy, reconstructed and repaired from per-meter data.\n* **Data quality**: every value we repaired or could not reconcile, with the portal's\n  original value.\n\nReference data is served from a local index synced from the portal's bulk export;\nreadings are fetched on demand and cached. `GET /v1/status` shows how fresh everything is.\nErrors use `application/problem+json` (RFC 9457) with a stable `code` (listed under the\n`Problem` schema). Unknown query parameters are rejected rather than ignored.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/transformers": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "List distribution transformers",
        "operationId": "list_transformers",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "feeder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only transformers on this feeder, e.g. `F-001`.",
              "title": "Feeder"
            },
            "description": "Only transformers on this feeder, e.g. `F-001`."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Transformer_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/insights/anomalies": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Fleet-wide anomaly report",
        "description": "Runs every anomaly rule over every cached series. Rules and severities:\ndata integrity (`duplicate_timestamps` info, `conflicting_duplicates`, `missing_values`,\n`gaps` warning), register behaviour (`register_decrease`, `implausible_load` error;\n`flatline`, `consumption_surge` warning; `power_factor_above_one` error), supply voltage\n(`no_voltage`, `voltage_outside_10pct` error; `voltage_outside_6pct` warning), status\n(`decommissioned_reporting` error) and freshness (`stale` warning: no reading for two\nintervals, at least two days).",
        "operationId": "anomaly_report",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rule",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "duplicate_timestamps",
                    "conflicting_duplicates",
                    "missing_values",
                    "gaps",
                    "register_decrease",
                    "implausible_load",
                    "flatline",
                    "consumption_surge",
                    "power_factor_above_one",
                    "no_voltage",
                    "voltage_outside_10pct",
                    "voltage_outside_6pct",
                    "decommissioned_reporting",
                    "stale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only list meters triggering this rule.",
              "title": "Rule"
            },
            "description": "Only list meters triggering this rule."
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "info",
                    "warning",
                    "error"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum severity to count and list.",
              "title": "Severity"
            },
            "description": "Minimum severity to count and list."
          },
          {
            "name": "include_meters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Set to false for just the counts.",
              "default": true,
              "title": "Include Meters"
            },
            "description": "Set to false for just the counts."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyReport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insights/consumption": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Consumption grouped by any dimension",
        "description": "Total energy per network node or meter attribute, over **complete IST days** only\n(a day needs a reading at both midnights). The window is widened to whole days. For\nexample: `group_by=transformer` to see transformer loading, `group_by=meter` for top\nconsumers, or `group_by=status` to see how much energy decommissioned meters use.",
        "operationId": "consumption_insight",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "zone",
                "circle",
                "division",
                "subdivision",
                "substation",
                "feeder",
                "transformer",
                "status",
                "make",
                "phase",
                "installation_type",
                "meter"
              ],
              "type": "string",
              "description": "A network level, a meter attribute, or `meter`.",
              "default": "transformer",
              "title": "Group By"
            },
            "description": "A network level, a meter attribute, or `meter`."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, `YYYY-MM-DD` (IST). Defaults to 7 days before `to`.",
              "title": "From"
            },
            "description": "First day, `YYYY-MM-DD` (IST). Defaults to 7 days before `to`."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, `YYYY-MM-DD`, inclusive (a datetime exactly at midnight closes the day before). Defaults to the latest cached reading.",
              "title": "To"
            },
            "description": "Last day, `YYYY-MM-DD`, inclusive (a datetime exactly at midnight closes the day before). Defaults to the latest cached reading."
          },
          {
            "name": "include_decommissioned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Decommissioned meters still report consumption; include them?",
              "default": true,
              "title": "Include Decommissioned"
            },
            "description": "Decommissioned meters still report consumption; include them?"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Return the top `limit` groups by consumption; see `groups_total`.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Return the top `limit` groups by consumption; see `groups_total`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumptionInsight"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnomalyReport": {
        "properties": {
          "meters_analysed": {
            "type": "integer",
            "title": "Meters Analysed",
            "description": "Meters whose readings are in the local cache."
          },
          "meters_total": {
            "type": "integer",
            "title": "Meters Total"
          },
          "meters_by_rule": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "duplicate_timestamps",
                "conflicting_duplicates",
                "missing_values",
                "gaps",
                "register_decrease",
                "implausible_load",
                "flatline",
                "consumption_surge",
                "power_factor_above_one",
                "no_voltage",
                "voltage_outside_10pct",
                "voltage_outside_6pct",
                "decommissioned_reporting",
                "stale"
              ]
            },
            "type": "object",
            "title": "Meters By Rule",
            "description": "How many meters trigger each rule."
          },
          "rule_severity": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "info",
                "warning",
                "error"
              ]
            },
            "propertyNames": {
              "enum": [
                "duplicate_timestamps",
                "conflicting_duplicates",
                "missing_values",
                "gaps",
                "register_decrease",
                "implausible_load",
                "flatline",
                "consumption_surge",
                "power_factor_above_one",
                "no_voltage",
                "voltage_outside_10pct",
                "voltage_outside_6pct",
                "decommissioned_reporting",
                "stale"
              ]
            },
            "type": "object",
            "title": "Rule Severity",
            "description": "Severity of each rule in `meters_by_rule`."
          },
          "meters_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "info",
                "warning",
                "error"
              ]
            },
            "type": "object",
            "title": "Meters By Severity",
            "description": "Distinct meters with at least one anomaly of each severity."
          },
          "meters": {
            "items": {
              "$ref": "#/components/schemas/MeterAnomalies"
            },
            "type": "array",
            "title": "Meters",
            "description": "Meters with anomalies. Meters whose only anomaly is `stale` are left out unless `rule=stale` is asked for; they still count in `meters_by_rule`."
          }
        },
        "type": "object",
        "required": [
          "meters_analysed",
          "meters_total",
          "meters_by_rule",
          "rule_severity",
          "meters_by_severity",
          "meters"
        ],
        "title": "AnomalyReport"
      },
...
      "ConsumptionInsight": {
        "properties": {
          "group_by": {
            "type": "string",
            "enum": [
              "zone",
              "circle",
              "division",
              "subdivision",
              "substation",
              "feeder",
              "transformer",
              "status",
              "make",
              "phase",
              "installation_type",
              "meter"
            ],
            "title": "Group By"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start",
            "description": "Midnight (IST) starting the first day of the window; null when there is no window because no readings are cached yet."
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End",
            "description": "Midnight (IST) after the last day of the window (exclusive)."
          },
          "days_counted": {
            "type": "integer",
            "title": "Days Counted",
            "description": "Distinct complete IST days that contributed to the totals."
          },
          "first_day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Day",
            "description": "First complete day counted."
          },
          "last_day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Day",
            "description": "Last complete day counted."
          },
          "include_decommissioned": {
            "type": "boolean",
            "title": "Include Decommissioned"
          },
          "meters_analysed": {
            "type": "integer",
            "title": "Meters Analysed",
            "description": "Meters with cached readings that were included."
          },
          "meters_total": {
            "type": "integer",
            "title": "Meters Total"
          },
          "total_kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Kwh",
            "description": "Null when no meter with cached readings was included."
          },
          "groups_total": {
            "type": "integer",
            "title": "Groups Total",
            "description": "Number of groups; `groups` holds the top `limit` by consumption."
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ConsumptionGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "group_by",
          "start",
          "end",
          "days_counted",
          "first_day",
          "last_day",
          "include_decommissioned",
          "meters_analysed",
          "meters_total",
          "total_kwh",
          "groups_total",
          "groups",
          "note"
        ],
        "title": "ConsumptionInsight"
      },
...
      "FieldError": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location",
            "description": "Where the value came from, e.g. `[\"query\", \"from\"]`."
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Machine-readable kind, e.g. `value_error` or `unknown_parameter`."
          }
        },
        "type": "object",
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "FieldError"
      },
...
      "MeterAnomalies": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "status": {
            "$ref": "#/components/schemas/MeterStatus"
          },
          "transformer_code": {
            "type": "string",
            "title": "Transformer Code"
          },
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyOut"
            },
            "type": "array",
            "title": "Anomalies"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "status",
          "transformer_code",
          "anomalies"
        ],
        "title": "MeterAnomalies"
      },
...
      "Page_Transformer_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Transformer"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of items matching the query (ignoring limit/offset)."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Transformer]"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code:\n\n* `unauthorized` (401): `X-API-Key` is missing or wrong (only when the server sets `URJA_API_KEY`).\n* `meter_not_found` (404): No meter with that id in the index, or the portal no longer has it.\n* `transformer_not_found` (404): No distribution transformer with that code.\n* `network_node_not_found` (404): No network node with that level and code.\n* `not_found` (404): No such path.\n* `method_not_allowed` (405): The path exists, but not for this HTTP method.\n* `http_error` (4xx): Any other HTTP-level error.\n* `validation_error` (422): A parameter is malformed, out of range or not recognised; `errors` says which.\n* `granularity_unavailable` (422): Hourly consumption asked of a meter that reports once a day.\n* `internal_error` (500): A bug in this service; the server log has the details.\n* `upstream_auth_failed` (502): This service could not log in to the portal.\n* `upstream_protocol_error` (502): The portal answered in a format this service does not understand.\n* `sync_rejected` (502): A sync returned data that looked broken; the previous snapshot is kept.\n* `index_not_ready` (503): The first sync with the portal hasn't finished; `Retry-After` gives the next attempt.\n* `upstream_unavailable` (503): The portal is down or too slow, and there is no cached copy to serve.\n* `upstream_rate_limited` (503): The portal kept answering 429, and there is no cached copy to serve.\n* `busy` (503): Too many requests are queued for the portal's rate budget.",
            "examples": [
              "meter_not_found"
            ]
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field-level validation errors."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details."
      },
...
      "Transformer": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "DT-001"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Malviya Nagar DT 1"
            ]
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code",
            "examples": [
              "F-001"
            ]
          },
          "capacity_kva": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kva",
            "examples": [
              100
            ]
          },
          "network": {
            "$ref": "#/components/schemas/NetworkPath",
            "description": "The transformer's own position in the network (zone \u2192 feeder)."
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "meters_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/MeterStatus"
            },
            "type": "object",
            "title": "Meters By Status"
          },
          "name_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Variants",
            "description": "Other names meters report for this transformer (stale data)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "network",
          "meter_count",
          "meters_by_status"
        ],
        "title": "Transformer"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Required only if the server is configured with URJA_API_KEY.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}